api/requirements.txt
Flask
flask-jwt-extended
python-dotenv
Werkzeug
